+++ MoltenVK/MoltenVK/OS/MTLSamplerDescriptor+MoltenVK.h
@@ -41,4 +41,13 @@
  */
 @property(nonatomic, readwrite) /*MTLSamplerBorderColor*/ NSUInteger borderColorMVK;
 
+/**
+ * Replacement for the lodBias property.
+ *
+ * This property allows support under all OS versions. Delegates to the lodBias
+ * property if it is available. otherwise, returns 0 when read and
+ * does nothing when set.
+ */
+@property(nonatomic, readwrite) float lodBiasMVK;
+
 @end