+++ MoltenVK/MoltenVK/Utility/MVKEnvironment.cpp
@@ -66,6 +66,9 @@ void mvkSetConfig(MVKConfiguration& dstMVKConfig, const MVKConfiguration& srcMVK
 	// Clamp timestampPeriodLowPassAlpha between 0.0 and 1.0.
 	dstMVKConfig.timestampPeriodLowPassAlpha = mvkClamp(dstMVKConfig.timestampPeriodLowPassAlpha, 0.0f, 1.0f);
 
+	// Only allow useMetalPrivateAPI to be enabled if we were built with support for it.
+	dstMVKConfig.useMetalPrivateAPI = dstMVKConfig.useMetalPrivateAPI && MVK_USE_METAL_PRIVATE_API;
+
 	// For each string member of the destination MVKConfiguration, store the contents
 	// in a std::string, then repoint the member to the contents of the std::string.
 #define MVK_CONFIG_MEMBER(member, mbrType, name)