+++ MoltenVK/MoltenVK/Vulkan/mvk_datatypes.mm
@@ -433,6 +433,52 @@ MVK_PUBLIC_SYMBOL MTLBlendFactor mvkMTLBlendFactorFromVkBlendFactor(VkBlendFacto
 	}
 }
 
+#if MVK_USE_METAL_PRIVATE_API
+
+// This isn't in any public header yet. I'm really just guessing based on the D3D11 values here.
+typedef NS_ENUM(NSUInteger, MTLLogicOperation) {
+	MTLLogicOperationClear,
+	MTLLogicOperationSet,
+	MTLLogicOperationCopy,
+	MTLLogicOperationCopyInverted,
+	MTLLogicOperationNoop,
+	MTLLogicOperationInvert,
+	MTLLogicOperationAnd,
+	MTLLogicOperationNand,
+	MTLLogicOperationOr,
+	MTLLogicOperationNor,
+	MTLLogicOperationXor,
+	MTLLogicOperationEquivalence,
+	MTLLogicOperationAndReverse,
+	MTLLogicOperationAndInverted,
+	MTLLogicOperationOrReverse,
+	MTLLogicOperationOrInverted,
+};
+
+MVK_PUBLIC_SYMBOL NSUInteger mvkMTLLogicOperationFromVkLogicOp(VkLogicOp vkLogicOp) {
+	switch (vkLogicOp) {
+		case VK_LOGIC_OP_CLEAR:			return MTLLogicOperationClear;
+		case VK_LOGIC_OP_AND:			return MTLLogicOperationAnd;
+		case VK_LOGIC_OP_AND_REVERSE:	return MTLLogicOperationAndReverse;
+		case VK_LOGIC_OP_COPY:			return MTLLogicOperationCopy;
+		case VK_LOGIC_OP_AND_INVERTED:	return MTLLogicOperationAndInverted;
+		case VK_LOGIC_OP_NO_OP:			return MTLLogicOperationNoop;
+		case VK_LOGIC_OP_XOR:			return MTLLogicOperationXor;
+		case VK_LOGIC_OP_OR:			return MTLLogicOperationOr;
+		case VK_LOGIC_OP_NOR:			return MTLLogicOperationNor;
+		case VK_LOGIC_OP_EQUIVALENT:	return MTLLogicOperationEquivalence;
+		case VK_LOGIC_OP_INVERT:		return MTLLogicOperationInvert;
+		case VK_LOGIC_OP_OR_REVERSE:	return MTLLogicOperationOrReverse;
+		case VK_LOGIC_OP_COPY_INVERTED:	return MTLLogicOperationCopyInverted;
+		case VK_LOGIC_OP_OR_INVERTED:	return MTLLogicOperationOrInverted;
+		case VK_LOGIC_OP_NAND:			return MTLLogicOperationNand;
+		case VK_LOGIC_OP_SET:			return MTLLogicOperationSet;
+		default:						return MTLLogicOperationCopy;
+	}
+}
+
+#endif
+
 MVK_PUBLIC_SYMBOL MTLVertexStepFunction mvkMTLVertexStepFunctionFromVkVertexInputRate(VkVertexInputRate vkVtxStep) {
 	switch (vkVtxStep) {
 		case VK_VERTEX_INPUT_RATE_VERTEX:		return MTLVertexStepFunctionPerVertex;